Pass []byte into UnmarshalJSON instead of io.Reader
During UnmarshalJSON, we need to parse the input twice. The first time
is to determine the geometry type, and the second time is to parse out
the coordinates. Before, the first pass also copied the input so that
the second pass could occur. Using a []byte instead for the input
eliminates this copy.