Make fuzz tests more robust
Fuzz tests were failing for two reasons. This change fixes the
failures by making the tests more robust:

1. GeoJSON comparisons were failing. This is because numerical
   differences are causing comparisons between 0 and very small negative
numbers. The comparison takes place using token comparison. However, the
negative sign of numbers was being parsed as a separate token compared
to the number itself, so we were getting a mismatch on the _number_ of
token in each GeoJSON.

2. For the boundary checks, we weren't using an XY tolerance. Some
   geometries were coming back with some slightly different digits in
the 14th and 15th decimal places. A tolerance parameter is now used to
account for this.