Remove IsValid
Having an IsValid method implies the existence of geometry objects that
were created in an invalid state. This is problematic, because all of
the methods on geometries assume that the geometry is valid.

Users will need alternate ways to determine if their WKB, WKT, or
GeoJSON are valid or not. This can be done by checking the error that
gets returned when the geometry is created.