Refactor geos package by unexporting Handle
The Handle type was originally unexported as an optimisation (the
intention was that a handle could be reused for multiple GEOS calls).
This was a premature optimisation (and turned out to not really make any
difference). It also produced an awkward situation where each method
(e.g. Equals) had to be documented in two places (on the Handle, and on
the global function). This resulted in a lot of duplicated comments.