Use uint64 rather than int for r-tree record ID
By using a uint64, we open up the possibility do do things other than
just storing an index to an array with the record id. E.g. we could
treat the uint64 as two uint32s, and store two different indexes. This
would be less clean with an int, because we can't be 100% guarenteed
that we have 64 bits to play with.