Make UnmarshalWKT accept a string instead of an io.Reader
This is because unmarshalling a WKT from a string seems to be the most
common operation. There is rarely a need to unmarshal from an io.Reader.
This also brings the UnmarshalWKT and UnmarshalWKB function signatures
in alignment with each other.

There is a new UnmarshalWKTFromReader function, which has the old
signature of UnmarshalWKT. This is only needed in those rare cases where
you have an io.Reader.