Use traditional pointer approach for RTree
The existing implementation uses an array/index based approach. There
was originally an intention that the data structure would be able to be
serialised to disk, hence an array/index based approach was used. That
seems a bit too specialised, and that approach was becoming very hard to
work with (e.g. to implement delete).

Using a more tradition node/pointer based approach allows easier
implementation of new extensions to the RTree (delete and ranged search
are the main ones in mind, but there may be more in the future).