Use direct passthrough for all GEOS operations
Previously, some GEOS operations such as Equals were defined in terms of
checking Relate masks. This required some preliminary checks to be
performed before hitting the main logic. These checks would now be
performed inside GEOS itself.

It is expected (although not confirmed) that there could be some
performance enhancements due to this change, because GEOS is now better
positioned to perform optimisations specific to each binary operation
that would not be possible when constructing a mask.

The original intention behind performing some of the logic in Go and
only wrapping Relate was so that we could gradually bring over _all_
operations to Go. That's not how the project evolved though (the Go-only
operations live entirely within the `geom` package).