Remove NewPointFromXY constructor
This constructor has too much redundancy with the NewPoint constructor.
The NewPoint contsructor is now supplemented by a new `AsPoint` helper
method on the `XY` type.

Note that the method signature of `AsPoint` and `NewPoint` will need to
change to additionally return an error as part of
https://github.com/peterstace/simplefeatures/issues/396 (so that will
need to get done before the next tag).