Rename constructors for consistency
The vast majority of constructors are now in the form `NewFoo`, where
`Foo` is the name of the thing being constructed (e.g. a geometry type
such as `MultiPolygon`).

Previously, the constructors were named so that the type that was being
constructed _from_ was also mentioned.

E.g. `NewPolygonFromRings` was renamed to `NewPolygon`. The `FromRings`
part really was redundent, since polygons are always constructed from
their constituent rings (there really is no other practical way to
construct them).