Re-write release notes for NaN/Inf validation
The original release notes had a lot of grammar mistakes in them, and
contained a lot of information in a single dot-point. Grammar problems
have been fixed, and explicit function/method signatures mentioned as
well. The notes have also been broken down into a few dot-points for
each part of the change.