Don't use explicit 'full' field in Envelope
Empty envelopes are now indicated by the "logical" value of Min X being
NaN. MinX is no longer stored directly as an envelope field, but instead
its bit patterned XORed with NaN is stored. This is so that the zero
value of an envelope has the logical value of Min X being NaN,
representing an empty envelope. This is in an effort to bring Envelope
back down to 32 bytes to see if that fixes the performance regression.