Fix "marshaler" spelling to match the Go std lib
It seems as though there are two spellings of "Marshal/Marshall" and
"Marshaller/Marshaller". When referring to the `encoding/json.Marshaler`
interface, we should use the same spelling as the standard library
(since they [define that
interface](https://pkg.go.dev/encoding/json#Marshaler)). For
consistency, we should use the same spelling in other places too.