Don't populate face labels for LineStrings
The original behaviour when constructing a DCEL from a LineString (or
MultiLineString) was to populate the face labels for half edges. They
were populated with the face being marked as "absent".

The assumption was that because LineStrings are not areal, there is
never a face for the that geometry in the final output. This was a valid
assumption because the input disallowed GeometryCollections (i.e. a mix
between areal and non-areal geometries).

If we allow GeometryCollection inputs, then the assumption is no longer
valid. This is because the face may need to be included in the geometry
by virtue of another (areal) part of the geometry collection (i.e. a
Polygon or a MultiPolygon).

The solution in this PR is to simply not populate face labels when
adding MultiLineStrings to a DCEL. If a Polygon is added to the DCEL as
a later step (as we're iterating over a GeometryCollection), then that
step will update the face labels accordingly.