Fix OmitInvalid bug for LineString and Polygon
The bug occurred when an invalid LineString or Polygon was constructed
using the OmitInvalid constructor option. It resulted in the Polygon or
LineString correctly being replaced with an EMPTY variant, however the
coordinates type was always set to DimXY.

The bugfix sets the coordinates type to match the input in the omit
invalid case.