Use Sequence instead of []XY for DCEL edges
The DCEL algorithm needs to represent the points along an edge. The
sources of these points are `Sequence` values from the input geometries.
The DCEL geometry extraction phase (the last step of the DCEL algorithm)
includes a step that inserts points along the edge into the output
geometry.

Before this change, DCEL edges represent points as a slice of `XY`. This
has the disadvantage of a conversion step from a `Sequence` to `[]XY`
(in the DCEL input stage), and then from `[]XY` back to `Sequence` (in
the DCEL output stage).

This change simplifies things by just using `Sequence` throughout. It
also slightly increases the range of points stored on each DCEL edge to
include the origin of the edge, and the final point on the next edge.
These points are redundant (since the output stage could read them from
the vertex records) but help with debugging. Because the Sequences are
shallow copies rather than deep copies, there is no performance penalty.