Add `UnaryUnion` and `UnionMany`
Because the DCEL algorithm supports `GeometryCollections`, the
`UnaryUnion` and `UnionMany` implementations are relatively trivial.

This PR also fixes a bug where an empty geometry and geometry collection
were input to `Union` or `(Symmetric)Difference`. The non-empty operand
was returned unaltered. The fixed behaviour now unions together the
non-empty geometry operand.