Add `BoundingDiagonal()` method to `Envelope`
The main usecase for this method is to help a user construct a compact
representation of an envelope, for example as WKB. Rather than encoding
the polygon representation of the Envelope, the LineString
representation can be used instead.

The Polygon representation uses 89 bytes (1 for the byte order, 4 to
encode the number of rings, 4 to encode the number of points in the
outer ring, and then 80 to represent the 5 points in the ring itself).
The LineString representation only uses 37 bytes (1 for the byte order,
4 to encode the number of points, and 32 to represent 2 points).