Simplify the `Type` method of the `Geometry` type
No need for the switch statement since we already have direct access to
the geometry type. The switch statement is likely a remnant of
extracting out geometry type (a _long_ time ago). The only functional
difference here is that if a Geometry is somehow constructed with an
invalid geometry type (which should be impossible), then the `Type`
method will no longer panic but instead just return the invalid geometry
type.