Simplify the bulk load algorithm
The previous bulk loading algorithm needed to keep the number of levels
consistent in the r-tree consistent (i.e. every leaf node had to be the
same distance from the root). This as a constraint required for the
`Insert` and `Delete` methods. Because those methods no longer exist, we
can greatly simplify the bulk loading algorithm.

With the simplified algorithm, the distance between the root and the
leaf nodes will differ by at most 1 level. Nodes are now packed much
more tightly, which reduces memory usage for some trees (but DOESN'T
improve CPU performance).