Remove `OmitInvalid` constructor option
This constructor isn't used in any non-trivial way by simplefeatures
itself, and isn't used by any known users.

It was introduced artificially to "show" that there are multiple uses
cases for constructor options (and thus the function options pattern was
warranted). However, the only truly used construction option is
`DisableAllValidations`.

The removal of `OmitInvalid` is the first step to change how geometries
are constructed without validation.