Remove constructor options from TransformXY methods
The `TransformXY` methods transforms the coordinates of geometries in
arbitrary ways via a user-supplied mapping. It should be obvious that
this potentially could cause invalid results. For example, the mapping
could trivially map every coordinate pair of a polygon to the same
point. In practice, transformations are either affine (translation,
rotation, scaling), or, at the very least, topologically preserving
(e.g. WGS84 to a Mercator or Conic projection).

This change removes the choice to validate or not in `TransformXY`
methods, always omitting a validation step. Users can still validate
manually via the `Validate` method if they wish to.