Simplify internal representation of Envelope
This change removes the "XOR NaN" hack. The hack was initially
introduced as a performance optimisation, but is fairly complicated.

NOTE: this PR introduces a performance regression of up to 5% for some
operations. The performance regression will be fixed in a subsequent PR
my modifying how the `Envelope` method for `LineString`s is implemented.