Upgrade golangci-lint to v1.55.2
Some of the changes involve the `revive` linter being a bit stricter,
resulting in the removal of "dot" imports. This is probably a good
thing, since the tests are now more similar to how users will see
things. There is also a new linter used called `perfsprint` which checks
for bad uses of `fmt.Sprintf`.