Hack to make line vs. line intersection symmetric
This fixes a bug where anomalously close DCEL nodes cause DCEL
extraction to fail. The bug would occur because the line vs. line
intersection operation was not symmetric, and was called from "both
directions" when re-noding input geometries.

This solution is a bit of a hack. Instead, a longer term solution is to
truly make the operation symmetric (e.g. by implementing a more
numerically stable algorithm).