Move low level details `geos` package into `internal/rawgeos`
This is to pave the way to allow unification of the GEOS wrapper between
the `geos` `cmprefimpl/cmpgeos` packages. The idea is that the
`internal/rawgeos` package will (soon) be a dependency of both packages,
with the extra wrapped functions that `cmpgeos` requires (soon) moved
from `cmpgeos` into `internal/rawageos`.