Allow the use of `pkg-config` to be opt-out
Some systems don't have `pkg-config`, or have `pkg-config` that's set up
incorrectly. These systems couldn't easily use the `v0.47.1` tag,
because that tag introduces a hard dependency on `pkg-config` being used
to configure GEOS.

This changeset allows `pkg-config` to be opted out of, by using the
`sfnopkgconfig` build tag. The behaviour when using that build tag is
effectively the same as pre-v0.47.1.