Retain coordinate type for empty simplified geometries
This fixes a bug where, in some scenarios, the `Simplify` operation
would drop extended coordinate type information (from XYZ, XYM, or XYZM
back to XY).

The bug happened because some geometry constructors inferred the
geometry type from the input's lowest common denominator coordinate
type. But if the input is empty, then the coordinate type is inferred as
XY. The constructor design is a bit of a sharp edge and is not addressed
here (instead, the bug that the sharp edge made more likely is fixed).