Add NewSingleRingPolygonXY[ZM] constructors
These fill the common usecase of creating a polygon with a single
(exterior) ring from a set of hardcoded coordinates.

	geom.NewPolygonXY([]float64{0, 0, 0, 4, 4, 4, 4, 0, 0, 0})

vs.

	geom.NewSingleRingPolygonXY(0, 0, 0, 4, 4, 4, 4, 0, 0, 0)

While the level of verbosity is about the same (similar character
count), the number of syntax items is reduced (no `[]float64{` or
closing `}`). This makes these cases more readable and easier to author.