Support XYZM in GeoJSON unmarshalling
The GeoJSON spec states:

> Implementations SHOULD NOT extend positions beyond three elements
> because the semantics of extra elements are unspecified and ambiguous.
> Historically, some implementations have used a fourth element to carry a
> linear referencing measure (sometimes denoted as "M") or a numerical
> timestamp, but in most situations a parser will not be able to properly
> interpret these values. The interpretation and meaning of additional
> elements is beyond the scope of this specification, and additional
> elements MAY be ignored by parsers.

In particular `additional elements MAY be ignored by parsers` indicates
that parsers may (or may not, in our case) choose to interpret
additional elements.

This changes interprets the 4th element as a measure (M) value.

Marshalling behaviour is unchanged, so the 4th element will not be
included in the marshalled GeoJSON output. This is required as per the
first sentence in the quote above.