Add support for GeoJSON Feature Collections foreign members
This change adds support for foreign members at the GeoJSON Feature
Collection level (foreign members are already supported at the GeoJSON
Feature level).

This necessitates a breaking change to the
`geom.GeoJSONFeatureCollection` type. It was previously defined as
`[]geom.GeoJSONFeature`, but is now defined as a struct with a `Features
[]geom.GeoJSONFeature` field and a `ForeignMembers`
map[string]json.RawMessage` field. The breaking change is necessary
because otherwise there is no where to put the foreign members.